// src/common/enums/domain-error-code.enum.ts
export enum DomainErrorCode {
  USER_SUSPENDED = 'USER_SUSPENDED',
  USER_ALREADY_ACTIVE = 'USER_ALREADY_ACTIVE',
  USER_ALREADY_INACTIVE = 'USER_ALREADY_INACTIVE',
  USER_ALREADY_SUSPENDED = 'USER_ALREADY_SUSPENDED',
  USER_NOT_SUSPENDED = 'USER_NOT_SUSPENDED',
  USER_ALREADY_EXISTS = 'USER_ALREADY_EXISTS',
  USER_NOT_FOUND = 'USER_NOT_FOUND',
  USER_ALREADY_HAS_ROLE = 'USER_ALREADY_HAS_ROLE',

  AUTH_IDENTITY_NOT_FOUND = 'AUTH_IDENTITY_NOT_FOUND',

  INVALID_PROVIDER = 'INVALID_PROVIDER',
  PROVIDER_CONFLICT = 'PROVIDER_CONFLICT',
  EMAIL_ALREADY_VERIFIED = 'EMAIL_ALREADY_VERIFIED',
  AUTH0_MANAGEMENT_ERROR = 'AUTH0_MANAGEMENT_ERROR',

  PROFILE_NOT_FOUND = 'PROFILE_NOT_FOUND',
  PROFILE_ALREADY_EXISTS = 'PROFILE_ALREADY_EXISTS',
  PHONE_ALREADY_IN_USE = 'PHONE_ALREADY_IN_USE',
  DNI_ALREADY_IN_USE = 'DNI_ALREADY_IN_USE',

  SERVICE_NOT_FOUND = 'SERVICE_NOT_FOUND',
  SERVICE_UNAUTHORIZED = 'SERVICE_UNAUTHORIZED',
  SERVICE_ALREADY_APPROVED = 'SERVICE_ALREADY_APPROVED',
  SERVICE_LIMIT_REACHED = 'SERVICE_LIMIT_REACHED',
  SERVICE_ALREADY_DISABLED = 'SERVICE_ALREADY_DISABLED',

  RATING_ALREADY_EXISTS = 'RATING_ALREADY_EXISTS',
  RATING_NOT_ALLOWED = 'RATING_NOT_ALLOWED',
  RATING_NOT_FOUND = 'RATING_NOT_FOUND',

  FAVORITE_ALREADY_EXISTS = 'FAVORITE_ALREADY_EXISTS',
  FAVORITE_NOT_FOUND = 'FAVORITE_NOT_FOUND',
  FAVORITE_NOT_ALLOWED = 'FAVORITE_NOT_ALLOWED',

  STRIPE_NO_ACTIVE_PRICE = 'STRIPE_NO_ACTIVE_PRICE',
  STRIPE_NO_CUSTOMER = 'STRIPE_NO_CUSTOMER',
  STRIPE_SESSION_FAILED = 'STRIPE_SESSION_FAILED',
  SUBSCRIPTION_ALREADY_ACTIVE = 'SUBSCRIPTION_ALREADY_ACTIVE',

  REPORT_NOT_FOUND = 'REPORT_NOT_FOUND',
  REPORT_COOLDOWN_ACTIVE = 'REPORT_COOLDOWN_ACTIVE',
  REPORT_TARGET_NOT_FOUND = 'REPORT_TARGET_NOT_FOUND',
}